feat: Major improvement for PIC+UART+VT100 behavior and minor MMIO API change...
We change our implementation of the UART Controller from the 16550D to
the 16750. The 16750 has support for hardware-control-flow (autoflow-control).
The VT100 Terminal now has an internal buffer and adheres to the UART control-flow.
This means, if the UART Controller's RX-FIFO reached it's trigger level, it will set
it's RTS to LOW, and the VT100 terminal stops sending. This in turn eliminates any
fifo-overrun errors on the UART Controller side and relieves the kernel.